+++ plugins/woocommerce-admin/packages/components/src/advanced-filters/number-filter.js
@@ -63,6 +63,7 @@ class NumberFilter extends Component {
 				components: {
 					filter: <Fragment>{ filterStr }</Fragment>,
 					rule: <Fragment>{ rule.label }</Fragment>,
+					title: <Fragment />,
 				},
 			} )
 		);