+++ plugins/woocommerce-admin/packages/components/src/advanced-filters/search-filter.js
@@ -91,6 +91,7 @@ class SearchFilter extends Component {
 				components: {
 					filter: <Fragment>{ filterStr }</Fragment>,
 					rule: <Fragment>{ rule.label }</Fragment>,
+					title: <Fragment />,
 				},
 			} )
 		);