+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 -   Fixed default value for `<Table />` component `onQueryChange` prop.Fixed default value for `<Table />` component `onQueryChange` prop.
 -   Deprecate our bespoke component `useFilters` in favor of using the WordPress variety `withFilters`.
+-   Fix screen reader text in `<AdvancedFilters />`.
 
 # 5.0.0